test: clarify that the assembly is for runtime tests
- move them to a Runtime folder
- append Runtime to assembly name
