+++ Eflatun.SceneReference/Assets/Tests/Runtime/ConvertedPathTests.cs.meta
