+++ Eflatun.SceneReference/Assets/Tests/Runtime/SceneReferenceUnitySerializationTests.cs
@@ -1,10 +1,10 @@
 ﻿using System.Collections;
-using Eflatun.SceneReference.Tests.Utils;
+using Eflatun.SceneReference.Tests.Runtime.Utils;
 using NUnit.Framework;
 using UnityEngine.SceneManagement;
 using UnityEngine.TestTools;
 
-namespace Eflatun.SceneReference.Tests
+namespace Eflatun.SceneReference.Tests.Runtime
 {
     public class SceneReferenceUnitySerializationTests
     {
