+++ Eflatun.SceneReference/Packages/com.eflatun.scenereference/Editor/AssemblyInfo.cs
@@ -1,8 +1,4 @@
-﻿using System.Runtime.CompilerServices;
-using UnityEngine.Scripting;
-
-// Make internals visible to the Test assembly
-[assembly: InternalsVisibleTo("Eflatun.SceneReference.Tests")]
+﻿using UnityEngine.Scripting;
 
 // Always link to make sure InitializeOnLoad methods don't get stripped away
 [assembly: AlwaysLinkAssembly]
