+++ Eflatun.SceneReference/Packages/com.eflatun.scenereference/Runtime/AssemblyInfo.cs
@@ -4,8 +4,8 @@ using UnityEngine.Scripting;
 // Make internals visible to the Editor assembly
 [assembly: InternalsVisibleTo("Eflatun.SceneReference.Editor")]
 
-// Make internals visible to the Test assembly
-[assembly: InternalsVisibleTo("Eflatun.SceneReference.Tests")]
+// Make internals visible to the runtime tests assembly
+[assembly: InternalsVisibleTo("Eflatun.SceneReference.Tests.Runtime")]
 
 // Always link to make sure RuntimeInitializeOnLoadMethod methods don't get stripped away
 [assembly: AlwaysLinkAssembly]
