+++ Eflatun.SceneReference/ProjectSettings/EditorBuildSettings.asset
@@ -6,13 +6,13 @@ EditorBuildSettings:
   serializedVersion: 2
   m_Scenes:
   - enabled: 0
-    path: Assets/Tests/Utils/TestScene_Disabled.unity
+    path: Assets/Tests/Runtime/Utils/TestScene_Disabled.unity
     guid: 7e37b14fa3517514a91937cec5cad27a
   - enabled: 1
-    path: Assets/Tests/Utils/TestScene_Enabled.unity
+    path: Assets/Tests/Runtime/Utils/TestScene_Enabled.unity
     guid: e3f2c1473b766c34ba5b37779d71787e
   - enabled: 1
-    path: Assets/Tests/Utils/TestScene_Container.unity
+    path: Assets/Tests/Runtime/Utils/TestScene_Container.unity
     guid: 4a619bd3d89b8d244944890d9aed7527
   - enabled: 1
     path: Assets/Demo/ValidationDemo.unity
